requirements.txt
Django==1.3.1
django-debug-toolbar==0.9.4
South==0.7.3

django-tastypie==0.9.11
requests==0.10.6

# For XML Support
lxml==2.3.3


# For the documentation
sphinx==1.1.2
